update open
some widgets like jq grid offer frozen columns which basically clones the selected columns and headers from the main layer and fixes them at a relative position to give the impression of being frozen. however this code was always referencing the button of the original layer that was scrollable. as a consequence the drop down would not appear where a user would expect it to in this particular use case. This code change address's the issue by always using the button that was used to show the select list.